refactor : LottoTicket 중복된 로직 수정 및 LottoTicekts 내부 요소를 합치는 로직 추가
