+++ src/test/java/lotto/domain/ticket/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
